Add files via upload
This version includes a filter tab in the sidebar that enables filtering functionality of experiments 1 and 2 using the same parameters with the click of a button. The genes identified are counted and displayed on the dash. 

The next update will add table visualization to the body of the data table tab.
Adding additional keyword parameters significantly increases the filters run time. I will examine the possibility to allow users to add more keywords if they want.